# Minimize tracing-tracy allocations