@SUPERCILEX 2023-12-27
Ok, this should be ready for review. I've updated the PR description with results.